fix setColorChooserColor

+++ u3-callbacks.h
@@ -17,5 +17,5 @@ void CarreChoixCouleurCB(Fl_Widget* w, void* data);
 void Slider1CB          (Fl_Widget* w, void* data);
 void Slider2CB          (Fl_Widget* w, void* data);
 void Slider3CB          (Fl_Widget* w, void* data);
-void setColorChooserColor(unsigned int couleur);
+void setColorChooserColor(unsigned long int couleur);
 #endif
